refactor: class 캡슐화 및 접근제어(final&private)
